feat(slide-in-motion): extract it to a separate component
